add "latemp" theme deps

+++ perl6-CI-Gen/lib/CI/Gen.pm6
@@ -132,18 +132,25 @@ EOF
 elif test "$cmd" = "before_install"
 then
 {$apt_str}
-    . /etc/lsb-release
-    sudo add-apt-repository -y ppa:inkscape.dev/stable
-    sudo apt -q update
-    sudo apt -y install inkscape
+    if test -e /etc/lsb-release
+    then
+        . /etc/lsb-release
+        sudo add-apt-repository -y ppa:inkscape.dev/stable
+        sudo apt -q update
+        sudo apt -y install inkscape
+    fi
+    cpanm --local-lib=~/perl_modules local::lib
     if test "$DISTRIB_ID" = 'Ubuntu'
     then
         if test "$DISTRIB_RELEASE" = '14.04'
         then
             sudo dpkg-divert --local --divert /usr/bin/ack --rename --add /usr/bin/ack-grep
         fi
+        eval "$(GIMME_GO_VERSION={self!calc-golang-version()} gimme)"
+    elif test -e /etc/fedora-release
+    then
+        sudo dnf --color=never install -y hspell-devel perl-devel ruby-devel
     fi
-    cpanm --local-lib=~/perl_modules local::lib
 {$before_install}
 EOF
 
@@ -225,6 +232,7 @@ END_OF_PROGRAM
 eval "$(GIMME_GO_VERSION={self!calc-golang-version()} gimme)"
 go get -u github.com/tdewolff/minify/cmd/minify
 {$local-lib-eval}
+cpanm -vvv IO::Async
 cpanm App::Deps::Verify App::XML::DocBook::Builder Pod::Xhtml
 {$install-T5-workaround}
 cpanm HTML::T5
@@ -234,7 +242,7 @@ cpanm --notest Bit::Vector Carp::Always Class::XSAccessor GD Getopt::Long IO::Al
 cpanm --notest Class::XSAccessor Config::IniFiles HTML::Links::Localize
 bash bin/install-git-cmakey-program-system-wide.bash 'git' 'src' 'https://github.com/thewml/website-meta-language.git'
 bash bin/install-git-cmakey-program-system-wide.bash 'git' 'installer' 'https://github.com/thewml/latemp.git'
-sudo -H `which python3` -m pip install beautifulsoup4 bs4 click cookiecutter lxml pycotap rebookmaker vnu_validator zenfilter Pillow WebTest
+sudo -H `which python3` -m pip install beautifulsoup4 bs4 click cookiecutter lxml pycotap rebookmaker vnu_validator weasyprint zenfilter Pillow WebTest
 perl bin/my-cookiecutter.pl
 # For various sites
 cpanm --notest HTML::Toc XML::Feed
@@ -251,7 +259,10 @@ bash bin/install-git-cmakey-program-system-wide.bash 'git' 'installer' 'https://
 pwd
 echo "HOME=$HOME"
 bash -x bin/install-npm-deps.sh
-sudo ln -s /usr/bin/make /usr/bin/gmake
+if ! test -e /usr/bin/gmake
+then
+    sudo ln -s /usr/bin/make /usr/bin/gmake
+fi
 END
             },
         );
